tests/api-resources/voice-did-management/voice-did-management.test.ts
// File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.

import Questblue from 'questblue';
import { Response } from 'node-fetch';

const client = new Questblue({
  username: 'My Username',
  password: 'My Password',
  securityKey: 'My Security Key',
  baseURL: process.env['TEST_API_BASE_URL'] ?? 'http://127.0.0.1:4010',
});

describe('resource voiceDidManagement', () => {
  test('fraudvalidate: only required params', async () => {
    const responsePromise = client.voiceDidManagement.fraudvalidate({ tn: 0 });
    const rawResponse = await responsePromise.asResponse();
    expect(rawResponse).toBeInstanceOf(Response);
    const response = await responsePromise;
    expect(response).not.toBeInstanceOf(Response);
    const dataAndResponse = await responsePromise.withResponse();
    expect(dataAndResponse.data).toBe(response);
    expect(dataAndResponse.response).toBe(rawResponse);
  });

  test('fraudvalidate: required and optional params', async () => {
    const response = await client.voiceDidManagement.fraudvalidate({ tn: 0 });
  });

  test('move2fax: only required params', async () => {
    const responsePromise = client.voiceDidManagement.move2fax({ did: 0 });
    const rawResponse = await responsePromise.asResponse();
    expect(rawResponse).toBeInstanceOf(Response);
    const response = await responsePromise;
    expect(response).not.toBeInstanceOf(Response);
    const dataAndResponse = await responsePromise.withResponse();
    expect(dataAndResponse.data).toBe(response);
    expect(dataAndResponse.response).toBe(rawResponse);
  });

  test('move2fax: required and optional params', async () => {
    const response = await client.voiceDidManagement.move2fax({ did: 0 });
  });
});
